First draft of app
Basic frontend UI. Added generateRatings.js, which generates the overall rating
and also returns the basic components making up the rating. Added a bunch of
edge case checking. The code is overall very messy.

Likely still lots of bugs that need to be fixed. Example, some artists were
not able to be queried using RYM search feature, resulting in a bug. I had
to check for this in the getRYMData code.